charset: change type in header, remove duplicate declaration in page.h

+++ include/charset.h
@@ -24,7 +24,7 @@
 #define __charset_h
 
 int char_digraph(int k1, int k2);
-int char_unicode_to_cp437(unsigned int c);
+uint8_t char_unicode_to_cp437(unsigned int c);
 uint8_t* str_unicode_to_cp437(const char* s);
 uint8_t* str_utf8_to_cp437(const char* s);
